+++ src/index.ts
@@ -6,6 +6,7 @@ import { ComponentEndpoint } from "./endpoints/component";
 import { ConfigurationEndpoint } from "./endpoints/configuration";
 import { CustomFieldEndpoint } from "./endpoints/customField";
 import { DashboardEndpoint } from "./endpoints/dashboard";
+import { EmailTemplatesEndpoint } from "./endpoints/emailTemplates";
 import { PermissionsEndpoint } from "./endpoints/permissions";
 import { Basic, BasicAuth } from "./types";
 
@@ -79,6 +80,10 @@ export class JiraServerConnector {
     dashboards: DashboardEndpoint;
 
     /**
+     * Contains all operations from '/rest/api/latest/email-templates'.
+     */
+    emailTemplates: EmailTemplatesEndpoint
+    /**
      * Instance new Jira Server Connector with user credentials and Jira host.
      * @param {BasicAuth} auth Basic Authorization info and Jira host 
      */
@@ -93,6 +98,7 @@ export class JiraServerConnector {
         this.configuration = new ConfigurationEndpoint(this.auth);
         this.customFields = new CustomFieldEndpoint(this.auth);
         this.dashboards = new DashboardEndpoint(this.auth);
+        this.emailTemplates = new EmailTemplatesEndpoint(this.auth);
     }